requirements.txt
# FastAPI関連
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# データベース関連
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.0

# 設定管理
pydantic-settings==2.0.3

# セキュリティ関連
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ユーティリティ関連
python-dotenv==1.0.0
qrcode[pil]==7.4.2
reportlab==4.0.7

# 開発・テスト関連
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2